Create set_state_cb for downloads and implemented a better download queue.
The state_cb is now set with a function called set_state_cb.  This then
removed the need to pass it as a parameter to the start methods of
download.

Using the state_cb better, the download list refreshes the queue on any
changes and will start any downloads that are queued when a download
slot opens up.